+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Yii 2 Application Runtime Configuration extension Change Log
 1.0.4 under development
 -----------------------
 
-- Enh #9: `StorageFilterTrait` created providing ability to use single entity for multiple configuration storage at `StorageDb`, `StorageMongoDb` and `StorageActiveRecord` (klimov-paul)
 - Enh: `Manager::configure()` allows to configure plain object not module only (klimov-paul)
+- Enh: Added `Manager::getItemValue()` allowing to get particular item value (klimov-paul)
+- Enh #9: `StorageFilterTrait` created providing ability to use single entity for multiple configuration storage at `StorageDb`, `StorageMongoDb` and `StorageActiveRecord` (klimov-paul)
 
 
 1.0.3, January 11, 2017